Modified the VerifyRefreshToken, and the VerifyUserCredentials middlewares. They were modified to store the users found in the database, in the req.databaseUser property. so that there is no need to query the database again, thus imporving performance.
After this modification, the database queries should be reduced by a half.
The routes wich make use of this middlewares were also modified to use the req.databaseUser property
instead of makeing a query to the database.

However, if something goes wrong, and the user found in the database is could
not be stored in the req.databaseUser property, the querie is still made to the database
but from the route.

We have to note that the both middlewares are never used in the same route.
Because in a route we can request a refresh token (for minor changes in the user's information),
or the user's credentials (for major changes in the user's information).
And to request the both would be redundant.